move orderedSet to ds pkg
